Normalize formatting and remove printf warnings
The source of this project is a total hack. Unified whitespacing is
a first step (no tabs, 4 spaces indent). Many of the xprintf:s had
problems with missing or extra arguments.

There was also warnings about %d for size_t. The 32/64 bit portable
solution is to use long format and cast to long, e.g "%lx" and
"(unsigned long)".