Introduce script to run c-xref in the background with visible, compact feedback
Since the tests for ffmpeg and systemd take a long time and we want compact
test output, this commit adds a script that runs the c-xref process in the
background and prints a "." each second for visible feedback.

It also adds check for errors in the output from the command.