Add discovery of standard include directories using cpp
Definitions in files included from the compilers standard directories goes
undetected. This is a specific problem if they change depending on other
options you use, such as language or ifdefs.

This change, from a patch by @litvindev, uses cpp to figure out where
the compiler finds those. Only works if the compiler is GCC-compatible.

This fixes #11, and fixes #8