Move .ok-files to .bs and build directly from them when BOOTSTRAPing
Previously we had .ok files that were known to work which could be
copied to the correspondig .h files to bootstrap. Now we conditionally
include the correct versions, which are the .bs files if we need to
bootstrap.

This means the actual strFill.h and strTdef.h should only exist for
local definitions and can be removed and added to .gitignore instead.