Bootstrap enumTxt by including generated enumTxt.h or enumTxt.bs
Replaced generating the enumTxt.c file directly and instead
generate a header file (somewhat in-correctly called because it
contains declarations, but since it is only included in enumTxt.c it
will do for now).

That include will either include the newly generated enumTxt.h (for
local adaptations) or the bootstrapping enumTxt.bs controlled by the
BOOTSTRAP preprocessor define.