Fix a bootstrapping problem introduced by init parsing
The initializer parsing changed the identifier struct which was
not reflected in the bootstrapping version of strFill.

How can we avoid this in the future? Since the strFill macros are
generated from the bootstrap version of c-xref I don't think we
can.... They must be hand-edited if any of the relevant structures
change.