Change naming strategy for generated and bootstrap files
In order to separate the header definitions for the enumTxt arrays
and its declarations we need both a generated .h and .c file.

So change naming to always end with a .c or .h and include a .g
for a generated file and and .bs for a bootstrap file.

Also introduce an enumTxt.h etc. which does the BOOTSTRAP magic, so
that any dependant can just #include "enumTxt.h" etc.