Rename a type marker field to something more than a single letter
This is a first example/test of renaming a field in a structure that
is subject to "fill macro generation" to see how hard that would be.

It turned out, as noted in NOTES.md that using c-xref itself fixed all
references except in the strFill.bs.h (obviously) that worked. So field
renames must be propagated to the "bs.h" files. But the compiler will tell
you where.