Allow specifying console and file levels separately
You can now specify which level to show on the console and in the log
file separately, so you can say "show errors on the console but
everything in the log file".

Also avoid printing the same message on the console twice if the file
pointer is stderr or stdout.