Replace test for position (ii) < 0 with assert()
In javaFQTypeSymbolDefinitionCreate() there was a test against
the argument (ii = position) being less than 0. This probably stems
from an earlier logic.

Currently I can see no way that this could happen, and there is no
coverage of that branch either, so I'm removing that and introducing
an assert() before the only call to that function that should catch
any such situation, should it arise in the future.