[parameters] Let the parser return another global "info-variable"
We need to know if the current parameter "list" actually is "void". In
that case when adding a parameter, the "void" should be removed.

In order to know this the C parser looks at the first symbol in the
"parameter_type_list" to see if that is "void". It then passes this to
the semantic function capturing parameter informtion which now
"returns" three variables: parameterListIsVoid, parameterBeginPosition
and parameterEndPosition.

With these we should be able to determine the correct replacement info
when adding or deleting from a void parameter list.