[parameters] Let the parser return parameter count
To be better able to verify if a parameter (nuumber) given is within
the "arity" of the function, e.g. add, delete, move parameter, the
parser now returns a parameter with the count.

This enabled such checks to be performed and errors issued, instead of
"could not parse parameter, continue?"