[refactor] Remove radically stupid block field in Memory struct
The Memory struct had a single double which was assumed could be
overflowed into _another_ variable that was supposed to be the actual
memory area. This was flake and would not be guaranteed to always
work.

As the only current usage if this is now in the options strings
allocation we can simply move the area, of correct size, into the
Memory struct itself.