[refactorings] Don't show Extract unless region is active
By not sending the mark unless the region is active we can be more
precise when figuring out which refactorings actually makes sense.

We should at some point also figure out how to know how many
parameters there are to a function so we can be more precise about
parameter refactorings such as not suggesting Delete Parameter when
there are none.