[java] Remove a lot of Java refactoring. Alas.
The Java refactoring in this project was pioneering work in the late
1990's. It probably represents decades of programming and software
research. There are more modern alternatives for Java but not so much
for C, so `c-xrefactory` will only support C/Yacc. So I'm saddly
removing all Java support including all this, and more to come. It
will still be available in the repo of course.