[macro] Fix collate() return value for token pasting
The previous commit introduced a bug where collate() would return
originalRightHandP, but this wasn't being updated correctly for all
code paths. When the RHS was a CPP_MACRO_ARGUMENT, it would point
to the argument token in the macro body (not advanced past it).

This fix properly tracks where to continue reading in the macro body:
- For macro arguments: skip past CPP_MACRO_ARGUMENT + extra info
- For regular operands: use the position after consuming the lexem

Fixes token pasting tests that were failing with syntax errors.